better description in README
Neat project, but was forced to run though a long scrolly landing page to figure out what exactly this is.  Homepage is well made & has fun animations, but many people decide in 0.3 seconds if a deeper dive is worth while; those who land on the Readme might not click on the URL. Hopefully the extra explainer paragraph will capture larger %.  (Also handy for newsletters to coppy/paste).