fix(react): `fetcher.load()` cancellations
The code was mistakenly cleaning up before it knew if it was aborted or not. By doing this, any aborted fetches would then clean up the controller for the fetch that canceled it, so the *next* one wouldn't find anything to cancel.

You can see it in the test before this commit

1. A fetches
2. B fetches, cancels A
3. A lands, cleans up B's abort controller (!)
4. C fetches, does not find B's controller because (3) cleaned it up prematurely

I knew it was off by one, didn't expect it to be off by one line number! 😂