fix: Remove maxAge from default session cookie
The options provided to `createCookieSessionStorage` are used as the default values of the session cookie. By setting `maxAge: 0` here, the session cookie will always be expired if the user doesn't override `maxAge` when they do `commitSession`. Leaving `maxAge` undefined will mean the cookie will expire when the browser session is closed. That is a reasonable default for a session cookie.

I believe this *bug* was originally in the Indie Stack that this tutorial was based on. The Stacks have since been fixed.