fix(remix-react): fix `<Form>` submit to not break `formMethod` functionality (#4053)
* test: failing test with `<Form>` not respecting `formMethod`

Bug report integration test demonstrating how the `<Form>` component –
unlike native `<form>` - does not respect the `formmethod` attribute set
on the submitter (the `<button>` submitting the form).

* fix(remix-react): submit `<Form>` w/method set by submitter `formmethod`

Fix `<Form>` component by not passing the form's `method` into the
`options` when calling the `submit(target, options)` function (created
with the `useSubmit()` hook). This ensures that the option does not take
precedence on any `formethod` set on the submitter which otherwise break
the `formmethod` functionality.

Since #3094 and #3094 the `<Form>` passes the "submitter" (if any) as
the `target` to the `submit(target, options)` (`useSubmitImpl`) which
will attempt to read the `formmethod`, `formaction` and `formenctype`
attributes on the target. Therefore, the responsability to infer which
method should be used should solely be on `useSubmitImpl` for the given
`target`.

Co-authored-by: Maxime Doury <maxime.doury@docavenue.com>

* readded tests from (#2162)

* Update integration/form-test.ts

Co-authored-by: Michaël De Boey <info@michaeldeboey.be>

* Update integration/form-test.ts

Co-authored-by: Michaël De Boey <info@michaeldeboey.be>

* Update integration/form-test.ts

Co-authored-by: Michaël De Boey <info@michaeldeboey.be>

* wrap two override tests in a describe

* consistent test text

* Create lovely-cats-jump.md

Co-authored-by: Nicholas Rakoto <nico@nrako.com>
Co-authored-by: Maxime Doury <maxime.doury@docavenue.com>
Co-authored-by: Michaël De Boey <info@michaeldeboey.be>
Co-authored-by: Logan McAnsh <logan@mcan.sh>